syle changes to user profile
